move PA feature for c sharp to scenario folder
