+++ samples/csharp/sharedcontent/console/Program.cs
@@ -37,7 +37,6 @@ namespace MicrosoftSpeechSDKSamples
                 Console.WriteLine(" 7. Standalone language detection samples.");
                 Console.WriteLine(" 8. Speech recognition with Microsoft Audio Stack (MAS) samples.");
                 Console.WriteLine(" 9. Diagnostics logging samples (trace logging).");
-                Console.WriteLine(" a. Pronunciation assessment samples.");
                 Console.WriteLine("");
                 Console.Write(mainPrompt);
 
@@ -81,9 +80,6 @@ namespace MicrosoftSpeechSDKSamples
                     case ConsoleKey.D9:
                     case ConsoleKey.NumPad9:
                         SpeechDiagnosticsLogging();
-                        break;
-                    case ConsoleKey.A:
-                        PronunciationAssessment();
                         break;
                     case ConsoleKey.D0:
                     case ConsoleKey.NumPad0:
@@ -788,71 +784,6 @@ namespace MicrosoftSpeechSDKSamples
                     case ConsoleKey.D5:
                     case ConsoleKey.NumPad5:
                         SpeechDiagnosticsLoggingSamples.MemoryLoggerWithOrWithoutFilter();
-                        break;
-                    case ConsoleKey.D0:
-                    case ConsoleKey.NumPad0:
-                        Console.WriteLine(back);
-                        sampleWasRun = false;
-                        break;
-                    default:
-                        Console.WriteLine(invalid);
-                        sampleWasRun = false;
-                        break;
-                }
-
-                if (sampleWasRun) Console.WriteLine(done);
-
-            } while (x.Key != ConsoleKey.D0);
-        }
-
-        //
-        // Shows pronunciation assessment samples.
-        // See more information at https://aka.ms/csspeech/pa
-        //
-        private static void PronunciationAssessment()
-        {
-            ConsoleKeyInfo x;
-
-            do
-            {
-                Console.WriteLine("");
-                Console.WriteLine(" Speech SDK - Pronunciation Assessment");
-                Console.WriteLine("");
-                Console.WriteLine(choose);
-                Console.WriteLine("");
-                Console.WriteLine(" 1. Pronunciation assessment with microphone input.");
-                Console.WriteLine(" 2. Pronunciation assessment with stream input.");
-                Console.WriteLine(" 3. Pronunciation assessment configured with json.");
-                Console.WriteLine(" 4. Pronunciation assessment continuous with file.");
-                Console.WriteLine(" 5. Pronunciation assessment with Microsoft Audio Stack.");
-                Console.WriteLine("");
-                Console.Write(prompt);
-
-                x = Console.ReadKey();
-                Console.WriteLine("\n");
-                bool sampleWasRun = true;
-
-                switch (x.Key)
-                {
-                    case ConsoleKey.D1:
-                    case ConsoleKey.NumPad1:
-                        SpeechRecognitionSamples.PronunciationAssessmentWithMicrophoneAsync().Wait();
-                        break;
-                    case ConsoleKey.D2:
-                    case ConsoleKey.NumPad2:
-                        SpeechRecognitionSamples.PronunciationAssessmentWithStream();
-                        break;
-                    case ConsoleKey.D3:
-                    case ConsoleKey.NumPad3:
-                        SpeechRecognitionSamples.PronunciationAssessmentConfiguredWithJson().Wait();
-                        break;
-                    case ConsoleKey.D4:
-                    case ConsoleKey.NumPad4:
-                        SpeechRecognitionSamples.PronunciationAssessmentContinuousWithFile().Wait();
-                        break;
-                    case ConsoleKey.D5:
-                    case ConsoleKey.NumPad5:
-                        SpeechRecognitionSamples.PronunciationAssessmentWithMas().Wait();
                         break;
                     case ConsoleKey.D0:
                     case ConsoleKey.NumPad0:
