+++ scenarios/csharp/dotnet-windows/language-learning/samples/packages.config
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Microsoft.CognitiveServices.Speech" version="1.45.0" targetFramework="net462" />
+  <package id="Microsoft.CognitiveServices.Speech.Extension.MAS" version="1.45.0" targetFramework="net462" />
+</packages>
