+++ scenarios/csharp/dotnetcore/language-learning/samples.sln
@@ -1,0 +1,24 @@
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.12.35728.132 d17.12
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "samples", "samples\samples.csproj", "{E1D0B282-8F55-43BC-B177-A2E1B8D455AE}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{E1D0B282-8F55-43BC-B177-A2E1B8D455AE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E1D0B282-8F55-43BC-B177-A2E1B8D455AE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E1D0B282-8F55-43BC-B177-A2E1B8D455AE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E1D0B282-8F55-43BC-B177-A2E1B8D455AE}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {D15FB198-E406-4BAD-9D5E-0C25DED6F5A3}
+	EndGlobalSection
+EndGlobal
