+++ scenarios/csharp/sharedcontent/console/README.md
@@ -1,0 +1,7 @@
+# Shared Content
+
+This directory contains shared content for two samples.
+Check out their documentation:
+
+* [C# Console app for .NET Framework on Windows](../../dotnet-windows/language-learning)
+* [C# Console app for .NET Core (Windows or Linux)](../../dotnetcore/language-learning)
