+++ scenarios/csharp/sharedcontent/console/zhcn_continuous_mode_sample.wav
